requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cssselect==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.32.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
kiwisolver==1.4.2
lxml==4.9.0
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pathtools==0.1.2
Pillow==9.1.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyparsing==3.0.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
torch
torchvision
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.20
pandas
